requirements.txt
pyglm
matplotlib
moderngl
numpy
Pillow
pygame
PyOpenGL
pywavefront
PyYAML
scipy
numpy-stl